Refactor some methods to scheduling library
