+++ apps/alarm/ChangeLog
@@ -19,3 +19,4 @@
 0.18: Cope with >1 identical alarm at once (#1667)
 0.19: Ensure rescheduled alarms that already fired have 'last' reset
 0.20: Use the new 'sched' factories to initialize new alarms/timers
+0.21: Refactor some methods to scheduling library
