+++ apps/sched/ChangeLog
@@ -3,3 +3,4 @@
 0.03: Fix `getTimeToAlarm` for a timer already used at same day, don't set `last` for timers. 
 0.04: Fix `getTimeToAlarm` to check for next dow if alarm.t lower currentTime.
 0.05: Export new functions (`newDefaultAlarm/Timer`), add Settings page
+0.06: Refactor some methods to library
