+++ apps/sched/lib.js
@@ -107,3 +107,20 @@ exports.getSettings = function () {
 exports.setSettings = function(settings) {
   require("Storage").writeJSON("sched.settings.json", settings);
 };
+
+// time in ms -> { hrs, mins }
+exports.decodeTime = function(t) {
+  t = 0 | t; // sanitise
+  let hrs = 0 | (t / 3600000);
+  return { hrs: hrs, mins: Math.round((t - hrs * 3600000) / 60000) };
+}
+
+// time in { hrs, mins } -> ms
+exports.encodeTime = function(o) {
+  return o.hrs * 3600000 + o.mins * 60000;
+}
+
+exports.formatTime = function(t) {
+  let o = decodeTime(t);
+  return o.hrs + ":" + ("0" + o.mins).substr(-2);
+}
