+++ apps/sched/metadata.json
@@ -1,7 +1,7 @@
 {
   "id": "sched",
   "name": "Scheduler",
-  "version": "0.05",
+  "version": "0.06",
   "description": "Scheduling library for alarms and timers",
   "icon": "app.png",
   "type": "scheduler",
