+++ apps/sleepphasealarm/ChangeLog
@@ -2,3 +2,4 @@
 0.02: Respect Quiet Mode
 0.03: Add compatibility for Bangle.js 2 and new firmware, added "Alarm at " for the alarm time
 0.04: Read alarms from new scheduling library, account for higher acceleration sensor noise on Bangle.js 2
+0.05: Refactor decodeTime() to scheduling library
