+++ apps/sleepphasealarm/metadata.json
@@ -2,11 +2,12 @@
   "id": "sleepphasealarm",
   "name": "SleepPhaseAlarm",
   "shortName": "SleepPhaseAlarm",
-  "version": "0.04",
+  "version": "0.05",
   "description": "Uses the accelerometer to estimate sleep and wake states with the principle of Estimation of Stationary Sleep-segments (ESS, see https://ubicomp.eti.uni-siegen.de/home/datasets/ichi14/index.html.en). This app will read the next alarm from the alarm application and will wake you up to 30 minutes early at the best guessed time when you are almost already awake.",
   "icon": "app.png",
   "tags": "alarm",
   "supports": ["BANGLEJS","BANGLEJS2"],
+  "dependencies": {"scheduler":"type"},
   "storage": [
     {"name":"sleepphasealarm.app.js","url":"app.js"},
     {"name":"sleepphasealarm.img","url":"app-icon.js","evaluate":true}
